+++ app/src/main/java/com/itg3/grp1/mobdevproject/DbTables/Listings.kt
@@ -53,7 +53,7 @@ class Listings(dbHandler: SQLiteOpenHelper) : IDbTable<Listing>(dbHandler)
         TODO("Not yet implemented")
     }
 
-    override fun update(instance: Listing): Long
+    override fun update(instance: Listing): Int
     {
         TODO("Not yet implemented")
     }
